prod-requirements.txt
django-twitter-bootstrap==3.0.0
django-admin-bootstrapped==1.6.1
South==0.8.3
pytz==2013.8
django-extensions==1.4.1
django-pipeline==1.3.16
#django-generic-flatblocks==0.9.1

#shop!